move get_courses_test to testing package and refactor to use mocks and helpers
